new fix

+++ preprocess_pipeline_crema.py
@@ -10,8 +10,7 @@ import pandas as pd
 import sklearn.metrics
 from sklearn.preprocessing import LabelEncoder
 import torch
-from transformers import Wav2Vec2ForSequenceClassification
-from torch.utils.data import DataLoader, TensorDataset
+
 
 def handle_MELD(directory):
     renaming_MELD_files.rename_audio_files(directory)
@@ -31,13 +30,13 @@ def handle_CREMA_D(directory):
         speaker.append(file.split("_")[0])
         line.append(file.split("_")[1])
         emotions.append(file.split("_")[2])
-        intensity.append(file.split("_")[3])
+        intensity.append(file.split("_")[3][:2])
 
     # Creating a csv file with the extracted labels
-    labels = pd.DataFrame(list(zip(index, speaker, line, emotions, intensity)), columns = ["Index", "filename", "Speaker", "Line", "Emotion", "Intensity"])
-    labels.to_csv(os.path.join(directory, "_labels.csv"), index=False)
+    labels = pd.DataFrame(list(zip(index, filename, speaker, line, emotions, intensity)), columns = ["Index", "filename", "Speaker", "Line", "Emotion", "Intensity"])
+    labels.to_csv(os.path.join(directory, "labels.csv"), index=False)
 
-    return os.path.join(directory, "_labels.csv")
+    return
 
 
 def handle_IEMOCAP(directory):
@@ -61,7 +60,12 @@ def switch_case(dataset, *args, **kwargs):
 
 def main():
     toggle_controls = [True, True, True, True]
-    dataset = "CREAMA-D"
+    dataset = "CREMA-D"
+    audio_directory = ""
+    corrected_labels_path = ""
+    labels_path = ""
+    video_directory = ""
+
 
     match dataset:
         case "MELD":
@@ -79,17 +83,19 @@ def main():
             audio_directory = r"C:\Users\DANIEL\Desktop\thesis\CREMA-D\audio"
 
             # Create the destination directory if it doesn't exist
-            os.makedirs(audio_directory, exist_ok=True)
+            if not os.path.exists(audio_directory):
+                os.makedirs(audio_directory, exist_ok=True)
 
-            # Copy all files from the source to the destination directory
-            for filename in os.listdir(old_audio_directory):
-                source_path = os.path.join(old_audio_directory, filename)
-                destination_path = os.path.join(audio_directory, filename)
-                shutil.copy2(source_path, destination_path)
+                # Copy all files from the source to the destination directory
+                for filename in os.listdir(old_audio_directory):
+                    source_path = os.path.join(old_audio_directory, filename)
+                    destination_path = os.path.join(audio_directory, filename)
+                    shutil.copy2(source_path, destination_path)
+
+
+            labels_path = os.path.join(os.path.dirname(audio_directory), "labels.csv")
 
             toggle_controls = [False, False, True, True]
-
-            labels_path = handle_CREMA_D(audio_directory)
             corrected_labels_path = os.path.join(os.path.dirname(labels_path), os.path.basename(labels_path)[:-4] + "_corrected.csv")
 
 
@@ -107,6 +113,7 @@ def main():
 
 
     if extract_audio_files_from_video:
+        print("Extracting audio files from video files...")
         if not os.path.exists(audio_directory):
             os.makedirs(audio_directory)
         if not os.listdir(audio_directory):
@@ -114,14 +121,22 @@ def main():
         else:
             print("Audio file folder not empty.")
 
+        print("Done.")
+
     if rename_files:
+        print("Renaming files...")
         switch_case(dataset, audio_directory)
+        print("Done.")
     
     if detect_broken_files:
+        print("Detecting broken audio files...")
         detect_broken_audio_files.process_files(audio_directory)
+        print("Done.")
     
     if extract_corrected_labels:
+        print("Extracting corrected labels...")
         match_labels.match_emotion_labels(labels_path,  corrected_labels_path, audio_directory)
+        print("Done.")
 
 if __name__ == "__main__":
     main()
